intrasafed-notpu/requirements.txt
Flask
pybase64
jsonpickle
Pillow
numpy
opencv-python
fastapi
